drivers/driver_ubx.c: clockbias and clockrate to GPSD JSON

+++ gpsd/gpsd_json.c
@@ -497,8 +497,9 @@ void json_tpv_dump(const gps_mask_t changed, struct gps_device_t *session,
     }
     if (0 != gpsdata->fix.clockbias &&
         0 != gpsdata->fix.clockdrift){
-        str_appendf(reply, replylen, ",\"clockbias\":%lu,\"clockdrift\":%lu",
-                    gpsdata->fix.clockbias, gpsdata->fix.clockdrift);
+        str_appendf(reply, replylen, ",\"clockbias\":%lld,\"clockdrift\":%lld",
+                    (long long)gpsdata->fix.clockbias,
+                    (long long)gpsdata->fix.clockdrift);
     }
     if (0 != (changed & NAVDATA_SET)) {
         if (0 != isfinite(gpsdata->fix.wanglem)){
